Compute features (#36)
Created files to add features to our dataframe. We compute features like
min, max, median, mean for each trip for a segment and also aggregates
for each of them. We refactored the script df_processing which computed
speeds and distances, to be part of compute_features folder. Thus, tests
were moved

---------

Co-authored-by: Lint Action <lint-action@samuelmeuli.com>
Co-authored-by: frejaz <fhpe19@student.aau.dk>
Co-authored-by: frejaz <56343992+frejaz@users.noreply.github.com>