Models (#47)
Made the code to create different models. For the regression model, a
"snap to closest speed limit" function has been implemented to be used
for those. Functionality for scoring our models is implemented. I used
metrics typically used in regression models, as I think those metrics
are more appropriate than those for classification. Much of the code is
not really typed, because it wraps code from sklearn, which does not
support type hints...

---------

Co-authored-by: Lint Action <lint-action@samuelmeuli.com>
Co-authored-by: frejaz <fhpe19@student.aau.dk>
Co-authored-by: frejaz <56343992+frejaz@users.noreply.github.com>