Fix array encoding (#84)
Array features should be encoded, should that there is a constant number
of elements in all the arrays of a columns (20 elements). They should be
split to 20 separate columns, such that there is 1 element in each row
of a new column. One hot encoding does not work. It creates more columns
than it should.

We decided not to use Vejman features. So encoding of categorical
features does not need to be fixed.
I changed many of the types from np.ndarray back to pandas dataframes.
As such, StatisticalModel should work.
Predictions made from our models are also pandas series, that can be
indexed with osm_id.

---------

Co-authored-by: CasperKjaerhus <casperkjaerhusleth@gmail.com>
Co-authored-by: Lint Action <lint-action@samuelmeuli.com>
Co-authored-by: anesheim-uni <57893968+anesheim-uni@users.noreply.github.com>
Co-authored-by: frejaz <fhpe19@student.aau.dk>